feat: blog management

+++ app/Http/Controllers/Api/Admin/BlogController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Helpers\ActivityHelpers;
 use App\Http\Helpers\APIHelpers;
 use App\Models\Blog;
+use App\Models\BlogComment;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -25,7 +26,7 @@ class BlogController extends Controller
             $data = Blog::with('author');
 
             if ($request->has('keyword')) {
-                $data = $data->where('nama_jenis_hewan', 'like', '%' . $request->keyword . '%');
+                $data = $data->where('judul', 'like', '%' . $request->keyword . '%');
             }
 
             if ($request->has('tahun')) {
@@ -42,7 +43,6 @@ class BlogController extends Controller
             return APIHelpers::responseAPI([
                 'data' => $data
             ], 200);
-
         } catch (Exception $error) {
             Log::error('Gagal mendapatkan data Blog!');
             ActivityHelpers::LogActivityHelpers('Gagal mendapatkan data Blog! (Admin)', ['message' => $error->getMessage()], '0');
@@ -55,10 +55,7 @@ class BlogController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
-    {
-        
-    }
+    public function create() {}
 
     /**
      * Store a newly created resource in storage.
@@ -70,7 +67,7 @@ class BlogController extends Controller
                 'judul' => 'required|unique:blogs',
                 'content' => 'required',
                 'tags' => 'required',
-                'thumbnail' => 'required|mimes:jpeg,jpg,png|max:2048',
+                'thumbnail_file' => 'required|mimes:jpeg,jpg,png|max:2048',
             ]);
 
             $data = [
@@ -83,11 +80,11 @@ class BlogController extends Controller
                 'slug' => Str::slug($validate['judul'])
             ];
 
-            if ($request->file('thumbnail')) {
-                $fileThumbnail = $validate['thumbnail']->store('/blog/thumbnail', 'public');
+            if ($request->file('thumbnail_file')) {
+                $fileThumbnail = $validate['thumbnail_file']->store('/blog/thumbnail', 'public');
                 $data['thumbnail'] = $fileThumbnail;
             }
-            
+
             Blog::create($data);
 
             ActivityHelpers::LogActivityHelpers('Membuat Blog', $data, '1');
@@ -107,9 +104,17 @@ class BlogController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(string $slug)
     {
-        //
+        try {
+            $data = Blog::with('komentar', 'author')->where('slug', $slug)->first();
+            Log::info('Berhasil mendapatkan detail Blog! (Admin)');
+            return APIHelpers::responseAPI($data, 200);
+        } catch (Exception $error) {
+            Log::error('Gagal mendapatkan detail Blog! (Admin)');
+            ActivityHelpers::LogActivityHelpers('Gagal mendapatkan detail Blog! (Admin)', ['message' => $error->getMessage()], '0');
+            return APIHelpers::responseAPI(['message' => $error->getMessage()], 500);
+        }
     }
 
     /**
@@ -123,13 +128,14 @@ class BlogController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, string $slug)
     {
         try {
             $validate = $request->validate([
                 'judul' => 'required',
                 'content' => 'required',
                 'tags' => 'required',
+                'thumbnail_file' => 'nullable|mimes:jpeg,jpg,png|max:2048',
             ]);
 
             $data = [
@@ -139,9 +145,26 @@ class BlogController extends Controller
                 'slug' => Str::slug($validate['judul'])
             ];
 
-            Blog::findOrFail($id)->update($data);
+            if ($request->file('thumbnail_file')) {
+                $fileThumbnail = $validate['thumbnail_file']->store('/blog/thumbnail', 'public');
+                $data['thumbnail'] = $fileThumbnail;
+            }
 
-            $data['id'] = $id;
+            $blog = Blog::where('slug', $slug)->first();
+
+            if (!$blog) {
+                ActivityHelpers::LogActivityHelpers('Gagal merubah Blog!', [
+                    'message' => 'Record dengan slug ' . $slug . 'tidak ditemukan'
+                ], '0');
+
+                return APIHelpers::responseAPI([
+                    'message' => 'Record dengan slug ' . $slug . 'tidak ditemukan'
+                ], 500);
+            }
+
+            $blog->update($data);
+
+            $data['id'] = $blog->id;
 
             ActivityHelpers::LogActivityHelpers('Mengedit Blog', $data, '1');
 
@@ -163,10 +186,24 @@ class BlogController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $slug)
     {
         try {
-            $data = Blog::findOrFail($id);
+            $data = Blog::where('slug', $slug)->first();
+
+            if (!$data) {
+                ActivityHelpers::LogActivityHelpers('Gagal merubah Blog!', [
+                    'message' => 'Record dengan slug ' . $slug . 'tidak ditemukan'
+                ], '0');
+
+                return APIHelpers::responseAPI([
+                    'message' => 'Record dengan slug ' . $slug . 'tidak ditemukan'
+                ], 500);
+            }
+            $commentBlog = BlogComment::where('blog_id', $data->id)->first();
+            if ($commentBlog) {
+                $commentBlog->delete();
+            }
             $data->delete();
             ActivityHelpers::LogActivityHelpers('Menghapus Blog', $data, '1');
             return APIHelpers::responseAPI($data, 200);
